requirements.txt
fastapi
uvicorn
openai
whisper
faiss-cpu
sentence-transformers
tts